hard-code SDS version properties
Note: this is a strange bug that should be tracked down!

For some reason this method:

  SdsConnect::Connect.properties

doesn't work when both the DIY and the SDS are run
from Phusion Passenger. Line 26 in lib/sds_connect.rb above should
look like this:

  @@config = @@config.merge(properties)

but that causes a hang on first page load.

If the statement SdsConnect::Connect.properties is placed in
config/initializers/diy.rb or app/controllers/application.rb
the startup process will hang.

However if it is placed in app/controlelrs/home_controller.rb
the startup process works normally.

So for now since all the SDSs we work with are version 1.1
these properties are hard-coded.